Add support for inspecting of hdf5's compound & time datatypes.
Note that this does NOT add reading or writing support. h5 will now
create DataSet objects for datasets with these types, but they cannot
be read into R itself.

This commit contains a change to the API in src/Helpers.h -- alloc
and read functions now need full H5::DataType information. This is not
strictly necessary for this feature, but it probably will be when full
support for e.g. compound datatypes is added.

Signed-off-by: Aaron Jacobs <atheriel@gmail.com>